Created exp_fetchers.py file.
