+++ core/controllers/admin.py
@@ -26,6 +26,7 @@ from core.domain import collection_services
 from core.domain import config_domain
 from core.domain import config_services
 from core.domain import exp_domain
+from core.domain import exp_fetchers
 from core.domain import exp_services
 from core.domain import recommendations_services
 from core.domain import rights_manager
@@ -276,7 +277,7 @@ class AdminHandler(base.BaseHandler):
             for i in range(num_dummy_exps_to_generate):
                 title = random.choice(possible_titles)
                 category = random.choice(constants.SEARCH_DROPDOWN_CATEGORIES)
-                new_exploration_id = exp_services.get_new_exploration_id()
+                new_exploration_id = exp_fetchers.get_new_exploration_id()
                 exploration = exp_domain.Exploration.create_default_exploration(
                     new_exploration_id, title=title, category=category,
                     objective='Dummy Objective')
@@ -392,7 +393,7 @@ class DataExtractionQueryHandler(base.BaseHandler):
         exp_id = self.request.get('exp_id')
         try:
             exp_version = int(self.request.get('exp_version'))
-            exploration = exp_services.get_exploration_by_id(
+            exploration = exp_fetchers.get_exploration_by_id(
                 exp_id, version=exp_version)
         except Exception:
             raise self.InvalidInputException(
