+++ core/controllers/library_test.py
@@ -23,6 +23,7 @@ from core.domain import activity_domain
 from core.domain import activity_services
 from core.domain import collection_services
 from core.domain import exp_domain
+from core.domain import exp_fetchers
 from core.domain import exp_jobs_one_off
 from core.domain import exp_services
 from core.domain import rating_services
@@ -235,7 +236,7 @@ class LibraryPageTests(test_utils.GenericTestBase):
     def test_library_handler_with_given_category_and_language_code(self):
         self.login(self.ADMIN_EMAIL)
 
-        exp_id = exp_services.get_new_exploration_id()
+        exp_id = exp_fetchers.get_new_exploration_id()
         self.save_new_valid_exploration(exp_id, self.admin_id)
         self.publish_exploration(self.admin_id, exp_id)
         exp_services.index_explorations_given_ids([exp_id])
