+++ core/domain/rating_services.py
@@ -19,6 +19,7 @@
 import datetime
 
 from core.domain import event_services
+from core.domain import exp_fetchers
 from core.domain import exp_services
 from core.platform import models
 import feconf
@@ -52,7 +53,7 @@ def assign_rating_to_exploration(user_id, exploration_id, new_rating):
         raise ValueError('Expected a rating 1-5, received %s.' % new_rating)
 
     try:
-        exp_services.get_exploration_by_id(exploration_id)
+        exp_fetchers.get_exploration_by_id(exploration_id)
     except:
         raise Exception('Invalid exploration id %s' % exploration_id)
 
@@ -74,7 +75,7 @@ def assign_rating_to_exploration(user_id, exploration_id, new_rating):
         return old_rating
     old_rating = transaction_services.run_in_transaction(_update_user_rating)
 
-    exploration_summary = exp_services.get_exploration_summary_by_id(
+    exploration_summary = exp_fetchers.get_exploration_summary_by_id(
         exploration_id)
     if not exploration_summary.ratings:
         exploration_summary.ratings = feconf.get_empty_ratings()
@@ -140,5 +141,5 @@ def get_overall_ratings_for_exploration(exploration_id):
         values are nonnegative integers representing the frequency counts
         of each rating.
     """
-    exp_summary = exp_services.get_exploration_summary_by_id(exploration_id)
+    exp_summary = exp_fetchers.get_exploration_summary_by_id(exploration_id)
     return exp_summary.ratings
