+++ core/domain/user_jobs_continuous.py
@@ -18,7 +18,7 @@ import ast
 import logging
 
 from core import jobs
-from core.domain import exp_services
+from core.domain import exp_fetchers
 from core.domain import feedback_services
 from core.domain import stats_services
 from core.platform import models
@@ -440,7 +440,7 @@ class UserStatsAggregator(jobs.BaseContinuousComputationManager):
                 model.total_plays += 1
                 model.put()
 
-        exp_summary = exp_services.get_exploration_summary_by_id(exp_id)
+        exp_summary = exp_fetchers.get_exploration_summary_by_id(exp_id)
         if exp_summary:
             for user_id in exp_summary.owner_ids:
                 if event_type == feconf.EVENT_TYPE_START_EXPLORATION:
@@ -589,7 +589,7 @@ class UserStatsMRJobManager(
         # Turn answer count into reach.
         reach = answer_count**exponent
 
-        exploration_summary = exp_services.get_exploration_summary_by_id(
+        exploration_summary = exp_fetchers.get_exploration_summary_by_id(
             item.id)
         contributors = exploration_summary.contributors_summary
         total_commits = sum(contributors.itervalues())
