+++ extensions/interactions/base_test.py
@@ -22,6 +22,7 @@ import string
 import struct
 
 from core.domain import dependency_registry
+from core.domain import exp_fetchers
 from core.domain import exp_services
 from core.domain import html_validation_service
 from core.domain import interaction_registry
@@ -581,7 +582,7 @@ class InteractionDemoExplorationUnitTests(test_utils.GenericTestBase):
             html_validation_service, 'get_filename_with_dimensions',
             mock_get_filename_with_dimensions):
             exp_services.load_demo(self._DEMO_EXPLORATION_ID)
-            exploration = exp_services.get_exploration_by_id(
+            exploration = exp_fetchers.get_exploration_by_id(
                 self._DEMO_EXPLORATION_ID)
 
         all_interaction_ids = set(
